+++ lightkurve/tests/test_targetpixelfile.py
@@ -9,7 +9,7 @@ import pytest
 import tempfile
 
 from ..targetpixelfile import KeplerTargetPixelFile, KeplerTargetPixelFileFactory
-from ..targetpixelfile import TessTargetPixelFile, open
+from ..targetpixelfile import TessTargetPixelFile
 from ..lightcurve import TessLightCurve
 from ..utils import LightkurveWarning
 
@@ -425,19 +425,3 @@ def test_tpf_slicing():
     assert tpf[5:10].shape == tpf.flux[5:10].shape
     assert tpf[0].targetid == tpf.targetid
     assert_array_equal(tpf[tpf.time < tpf.time[5]].time, tpf.time[0:5])
-
-
-def test_open():
-    k2_path = os.path.dirname(os.path.abspath(__file__)) + '/data/test-tpf-star.fits'
-    tess_path = os.path.dirname(os.path.abspath(__file__)) + '/data/tess25155310-s01-first-cadences.fits.gz'
-    k2tpf = open(k2_path)
-    assert isinstance(k2tpf, KeplerTargetPixelFile)
-    tesstpf = open(tess_path)
-    assert isinstance(tesstpf, TessTargetPixelFile)
-    try:
-        open(os.path.dirname(os.path.abspath(__file__)) +
-                  '/data/test_factory0.fits')
-    except ValueError:
-        pass
-    assert isinstance(KeplerTargetPixelFile(tess_path), KeplerTargetPixelFile)
-    assert isinstance(TessTargetPixelFile(k2_path), TessTargetPixelFile)
